jitendra - logs function and favicon icon added
What changes have you made?

insert the favicon icon file in public and wrote a function which will
print the all logs of the server

Why do we need this change?

it would be nice if we will able to track all the logs of the server

What can this change impact?
	new file:   .DS_Store
	new file:   Tpdkdesign.net-Refresh-Cl-Symbols-Tips.ico
	modified:   lib/controller.js
	modified:   public/.DS_Store
	modified:   public/css/.DS_Store
	modified:   public/favicon.ico
	modified:   server.js

 Changes not staged for commit:
	modified:   server.js